Issue Updated - Multiple ids with the same value
When I click on one input it always points to name field.
In the starting structure provided, the id field for all 3 input elements is the same. I guess this is unintentional typo and not a best practice.
Please merge it so other people like don't get confuse, great course as always.