Update Modal.css
When the model is closed we should set its z-index to some negative value since we're not displaying it none. Because with opacity the element is still there and it might overlap with other elements or a button. In my case, the element was above the "open model" button so I wasn't able to click it. This change will take care of that issue. :)